// src/core/types/meta-catalog/meta-property-feed.types.ts
export interface MetaPropertyFeedItem {
  home_listing_id: string;
  name: string;
  description: string;
  availability: string;
  price: string;
  property_type: string;
  garden_type: string;
  url: string;
  num_baths: number | null;
  num_beds: number | null;
  num_rooms: number | null;
  pet_policy: string;
  area_size: number | null;
  land_area_size: number | null;
  parking_spaces: number | null;
  area_unit: string;
  year_built: string;
  "address.addr1": string;
  "address.city": string;
  "address.region": string;
  "address.country": string;
  "neighborhood[0]": string;
  virtual_tour_url: string;
  "image[0].url": string;
  "image[0].tag[0]": string;
  "image[1].url": string;
  "image[1].tag[0]": string;
  "image[2].url": string;
  "image[2].tag[0]": string;
  "image[3].url": string;
  "image[3].tag[0]": string;
  "image[4].url": string;
  "image[4].tag[0]": string;
  "image[5].url": string;
  "image[5].tag[0]": string;
  "image[6].url": string;
  "image[6].tag[0]": string;
  "image[7].url": string;
  "image[7].tag[0]": string;
  [key: string]: string | number | null;
}